Radically altered how the analysis is performed: instead of visiting a class structure (asm style) just iterate over the constant pool of the class, and raise a bug if a deprecated type appears somewhere.
Bug reporting will suffer in that bugs will not point to specific uses (e.g. is deprecated type referenced as a field, parameter, in generics, etc?) however, the accuracy of the analysis should benefit, since this method should find *any* case where a deprecated type is referenced, and I won't miss edge cases (e.g. generics in an anonymous instance). Hopefully this is the better choice.